alarm/test_alarm_permitted_app.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Permitted Application for Alarm API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript">

SpecialPowers.addPermission("alarms", true, document);
SpecialPowers.addPermission("alarm", true, document);

SimpleTest.waitForExplicitFinish();

SpecialPowers.pushPrefEnv({"set": [["dom.mozAlarms.enabled", true]]}, function() {
 
  // mozAlarms is intalled on all platforms except Android for the moment.
  ok('mozAlarms' in navigator, "navigator.mozAlarms should exist");
  ok(navigator.mozAlarms, "navigator.mozAlarms should return an object");
  ok(navigator.mozAlarms instanceof Components.interfaces.nsIDOMMozAlarmsManager,
    "navigator.mozAlarms should be an nsIDOMAlarmsManager object");

  SpecialPowers.removePermission("alarms", document);
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
